+++ Commands/Provisioning/Site/SaveProvisioningTemplate.cs
@@ -3,6 +3,7 @@ using OfficeDevPnP.Core.Framework.Provisioning.Model;
 using OfficeDevPnP.Core.Framework.Provisioning.Providers;
 using OfficeDevPnP.Core.Framework.Provisioning.Providers.Xml;
 using SharePointPnP.PowerShell.CmdletHelpAttributes;
+using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 using SharePointPnP.PowerShell.Commands.Utilities;
 using System;
 using System.Collections.Generic;
@@ -18,17 +19,34 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
     [CmdletHelp("Saves a PnP site template to the file system",
         Category = CmdletHelpCategory.Provisioning)]
     [CmdletExample(
-       Code = @"PS:> Save-PnPSiteTemplate -InputInstance $template -Out .\template.pnp",
-       Remarks = "Saves a PnP site template to the file system as a PnP file.",
+       Code = @"PS:> Save-PnPProvisioningTemplate -Template .\template.xml -Out .\template.pnp",
+       Remarks = "Saves a PnP provisioning template to the file system as a PnP file.",
        SortOrder = 1)]
+    [CmdletExample(
+       Code = @"PS:> $template = Read-PnPProvisioningTemplate -Path template.xml
+PS:> Save-PnPProvisioningTemplate -Template $template -Out .\template.pnp",
+       Remarks = "Saves a PnP site template to the file system as a PnP file. The schema used will the latest released schema when creating the PnP file regardless of the original schema",
+       SortOrder = 2)]
+    [CmdletExample(
+       Code = @"PS:> $template = Read-PnPProvisioningTemplate -Path template.xml
+PS:> Save-PnPProvisioningTemplate -Template $template -Out .\template.pnp -Schema V202002",
+       Remarks = "Saves a PnP site template to the file system as a PnP file  and converts the template in the PnP file to the specified schema.",
+       SortOrder = 3)]
+    [CmdletExample(
+       Code = @"PS:> Read-PnPProvisioningTemplate -Path template.xml | Save-PnPProvisioningTemplate -Out .\template.pnp",
+       Remarks = "Saves a PnP site template to the file system as a PnP file.",
+       SortOrder = 4)]
     public class SaveProvisioningTemplate : PSCmdlet
     {
-        [Parameter(Mandatory = true, HelpMessage = "Allows you to provide an in-memory instance of the ProvisioningTemplate type of the PnP Core Component. When using this parameter, the -Out parameter refers to the path for saving the template and storing any supporting file for the template.")]
+        [Parameter(Mandatory = true, ValueFromPipeline = true, HelpMessage = "Allows you to provide an in-memory instance of the ProvisioningTemplate type of the PnP Core Component. When using this parameter, the -Out parameter refers to the path for saving the template and storing any supporting file for the template.")]
         [Alias("InputInstance")]
-        public ProvisioningTemplate Template;
+        public ProvisioningTemplatePipeBind Template;
 
         [Parameter(Mandatory = true, Position = 0, HelpMessage = "Filename to write to, optionally including full path.")]
         public string Out;
+
+        [Parameter(Mandatory = false, HelpMessage = "The optional schema to use when creating the PnP file. Always defaults to the latest schema.")]
+        public XMLPnPSchemaVersion Schema = XMLPnPSchemaVersion.LATEST;
 
         [Parameter(Mandatory = false, HelpMessage = "Specifying the Force parameter will skip the confirmation question.")]
         public SwitchParameter Force;
@@ -38,6 +56,11 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
 
         protected override void ProcessRecord()
         {
+            var templateObject = Template.GetTemplate(SessionState.Path.CurrentFileSystemLocation.Path, (e) =>
+            {
+                WriteError(new ErrorRecord(e, "TEMPLATENOTVALID", ErrorCategory.SyntaxError, null));
+            });
+
             // Determine the output file name and path
             string outFileName = Path.GetFileName(Out);
 
@@ -79,7 +102,7 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
 
             var fileSystemConnector = new FileSystemConnector(outPath, "");
 
-            ITemplateFormatter formatter = XMLPnPSchemaFormatter.LatestFormatter;
+            ITemplateFormatter formatter = ProvisioningHelper.GetFormatter(Schema);
 
             if (extension == ".pnp")
             {
@@ -90,64 +113,64 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
                       Out, fileSystemConnector);
                 var templateFileName = outFileName.Substring(0, outFileName.LastIndexOf(".", StringComparison.Ordinal)) + ".xml";
 
-                provider.SaveAs(Template, templateFileName, formatter, TemplateProviderExtensions);
-                ProcessFiles(Out, fileSystemConnector, provider.Connector);
+                provider.SaveAs(templateObject, templateFileName, formatter, TemplateProviderExtensions);
+                ProcessFiles(templateObject, Out, fileSystemConnector, provider.Connector);
             }
             else
             {
                 XMLTemplateProvider provider = new XMLFileSystemTemplateProvider(outPath, "");
-                provider.SaveAs(Template, Out, formatter, TemplateProviderExtensions);
+                provider.SaveAs(templateObject, Out, formatter, TemplateProviderExtensions);
             }
         }
 
-        private void ProcessFiles(string templateFileName, FileConnectorBase fileSystemConnector, FileConnectorBase connector)
+        private void ProcessFiles(ProvisioningTemplate template, string templateFileName, FileConnectorBase fileSystemConnector, FileConnectorBase connector)
         {
             var templateFile = ReadProvisioningTemplate.LoadProvisioningTemplateFromFile(templateFileName, null, (e) =>
             {
                 WriteError(new ErrorRecord(e, "TEMPLATENOTVALID", ErrorCategory.SyntaxError, null));
             });
-            if (Template.Tenant?.AppCatalog != null)
+            if (template.Tenant?.AppCatalog != null)
             {
-                foreach (var app in Template.Tenant.AppCatalog.Packages)
+                foreach (var app in template.Tenant.AppCatalog.Packages)
                 {
                     WriteObject($"Processing {app.Src}");
                     AddFile(app.Src, templateFile, fileSystemConnector, connector);
                 }
             }
-            if (Template.Tenant?.SiteScripts != null)
+            if (template.Tenant?.SiteScripts != null)
             {
-                foreach (var siteScript in Template.Tenant.SiteScripts)
+                foreach (var siteScript in template.Tenant.SiteScripts)
                 {
                     WriteObject($"Processing {siteScript.JsonFilePath}");
                     AddFile(siteScript.JsonFilePath, templateFile, fileSystemConnector, connector);
                 }
             }
-            if (Template.Localizations != null && Template.Localizations.Any())
+            if (template.Localizations != null && template.Localizations.Any())
             {
-                foreach (var location in Template.Localizations)
+                foreach (var location in template.Localizations)
                 {
                     WriteObject($"Processing {location.ResourceFile}");
                     AddFile(location.ResourceFile, templateFile, fileSystemConnector, connector);
                 }
             }
 
-            if (Template.WebSettings != null && !String.IsNullOrEmpty(Template.WebSettings.SiteLogo))
+            if (template.WebSettings != null && !String.IsNullOrEmpty(template.WebSettings.SiteLogo))
             {
                 // is it a file?
                 var isFile = false;
-                using (var fileStream = fileSystemConnector.GetFileStream(Template.WebSettings.SiteLogo))
+                using (var fileStream = fileSystemConnector.GetFileStream(template.WebSettings.SiteLogo))
                 {
                     isFile = fileStream != null;
                 }
                 if (isFile)
                 {
-                    WriteObject($"Processing {Template.WebSettings.SiteLogo}");
-                    AddFile(Template.WebSettings.SiteLogo, templateFile, fileSystemConnector, connector);
+                    WriteObject($"Processing {template.WebSettings.SiteLogo}");
+                    AddFile(template.WebSettings.SiteLogo, templateFile, fileSystemConnector, connector);
                 }
             }
-            if (Template.Files.Any())
+            if (template.Files.Any())
             {
-                foreach (var file in Template.Files)
+                foreach (var file in template.Files)
                 {
                     WriteObject($"Processing {file.Src}");
                     AddFile(file.Src, templateFile, fileSystemConnector, connector);